Revert "added contact and WebView cookies access"
