+++ MANIFEST
@@ -1,0 +1,4 @@
+# file GENERATED by distutils, do NOT edit
+__init__.py
+kvdroid.py
+setup.py
