+++ kvdroid/metrics.py
@@ -1,8 +1,5 @@
+from kvdroid import activity, Point
 import re
-
-from pyandroidkx import activity
-from pyandroidkx.jclass.android.graphics import Point
-
 
 class Metrics(object):
     config = activity.getResources().getConfiguration()
@@ -25,13 +22,11 @@ class Metrics(object):
             return "portrait"
         else:
             return "landscape"
-
-    @staticmethod
-    def resolution():
+            
+    def resolution(self):
         point = Point()
         activity.getWindowManager().getDefaultDisplay().getRealSize(point)
-        size = re.findall("\d+", point.toString())
-        return size[1] + "x" + size[0]
-
+        size = re.findall("\d+",point.toString())
+        return size[1]+"x"+size[0]
 
 screen = Metrics()
