--- pyandroidkx/jclass/android/content/__init__.py
@@ -1,5 +1,0 @@
-from jnius import autoclass
-
-Intent = autoclass('android.content.Intent')
-Context = autoclass('android.content.Context')
-IntentFilter = autoclass("android.content.IntentFilter")
