--- pyandroidkx/jclass/android/content/res.py
@@ -1,3 +1,0 @@
-from jnius import autoclass
-
-Configuration = autoclass("android.content.res.Configuration")
