--- pyandroidkx/jclass/android/graphics.py
@@ -1,6 +1,0 @@
-from jnius import autoclass
-
-Color = autoclass("android.graphics.Color")
-Rect = autoclass('android.graphics.Rect')
-BitmapFactory = autoclass('android.graphics.BitmapFactory')
-Point = autoclass("android.graphics.Point")
