--- pyandroidkx/jclass/android/media.py
@@ -1,4 +1,0 @@
-from jnius import autoclass
-
-MediaPlayer = autoclass('android.media.MediaPlayer')
-AudioManager = autoclass('android.media.AudioManager')
