--- pyandroidkx/jclass/android/net.py
@@ -1,4 +1,0 @@
-from jnius import autoclass
-
-Uri = autoclass('android.net.Uri')
-ConnectivityManager = autoclass('android.net.ConnectivityManager')
