--- pyandroidkx/jclass/android/provider.py
@@ -1,5 +1,0 @@
-from jnius import autoclass
-
-Provider = autoclass('android.provider.Settings')
-Contacts = autoclass('android.provider.ContactsContract$Contacts')
-Phone = autoclass('android.provider.ContactsContract$CommonDataKinds$Phone')
