--- pyandroidkx/jclass/android/speech/tts.py
@@ -1,3 +1,0 @@
-from jnius import autoclass
-
-TextToSpeech = autoclass('android.speech.tts.TextToSpeech')
