--- pyandroidkx/jclass/android/webkit.py
@@ -1,3 +1,0 @@
-from jnius import autoclass
-
-CookieManager = autoclass("android.webkit.CookieManager")
