--- pyandroidkx/jclass/android/widget.py
@@ -1,3 +1,0 @@
-from jnius import autoclass
-
-Toast = autoclass('android.widget.Toast')
