--- pyandroidkx/jclass/androidx/__init__.py
@@ -1,8 +1,0 @@
-from jnius import autoclass, JavaException
-
-
-try:
-    autoclass("androidx.core.app.ActivityCompat")
-except JavaException as e:
-    raise JavaException(
-        f"{e}\nadd androidx.appcompat:appcompat:1.2.0 to buildozer.spec file: android.gradle_dependencies")
