--- pyandroidkx/jclass/androidx/core/content/__init__.py
@@ -1,3 +1,0 @@
-from jnius import autoclass
-
-ContextCompat = autoclass("androidx.core.content.ContextCompat")
