--- pyandroidkx/jclass/java/io.py
@@ -1,3 +1,0 @@
-from jnius import autoclass
-
-File = autoclass('java.io.File')
