--- pyandroidkx/jclass/java/lang.py
@@ -1,5 +1,0 @@
-from jnius import autoclass
-
-Runtime = autoclass('java.lang.Runtime')
-String = autoclass("java.lang.String")
-System = autoclass("java.lang.System")
