--- pyandroidkx/jclass/java/net.py
@@ -1,5 +1,0 @@
-from jnius import autoclass
-
-URLConnection = autoclass("java.net.URLConnection")
-HttpURLConnection = autoclass("java.net.HttpURLConnection")
-URL = autoclass("java.net.URL")
