--- pyandroidkx/jclass/java/util.py
@@ -1,4 +1,0 @@
-from jnius import autoclass
-
-Locale = autoclass('java.util.Locale')
-ArrayList = autoclass('java.util.ArrayList')
