--- pyandroidkx/jimplement/customtabs.py
@@ -1,22 +1,0 @@
-def launch_url(url: str, color="#FFFFFF", color_scheme="system"):
-    from pyandroidkx.jclass.android.graphics import Color
-    from pyandroidkx.jclass.androidx.browser.customtabs import CustomTabColorSchemeParamsBuilder
-    from pyandroidkx.jclass.androidx.browser.customtabs import CustomTabsIntentBuilder, CustomTabsIntent
-    from pyandroidkx import activity
-    from pyandroidkx.jclass.android.net import Uri
-    schemes = {
-        "system": CustomTabsIntent.COLOR_SCHEME_SYSTEM,
-        "light": CustomTabsIntent.COLOR_SCHEME_LIGHT,
-        "dark": CustomTabsIntent.COLOR_SCHEME_DARK
-    }
-    color_int = Color.parseColor(color)
-    default_color = CustomTabColorSchemeParamsBuilder().setToolbarColor(color_int).build()
-    builder = CustomTabsIntentBuilder()
-    if color_scheme not in schemes:
-        builder.setColorScheme(CustomTabsIntent.COLOR_SCHEME_SYSTEM)
-    else:
-        builder.setColorScheme(schemes[color_scheme])
-    builder.setDefaultColorSchemeParams(default_color)
-    custom_tabs_intent = builder.build()
-    custom_tabs_intent.intent.setPackage("com.android.chrome")
-    custom_tabs_intent.launchUrl(activity, Uri.parse(url))
