+++ setup.py
@@ -1,4 +1,4 @@
-from setuptools import setup, find_packages
+from setuptools import setup
 
 from os import path
 
@@ -7,19 +7,16 @@ with open(path.join(this_directory, 'README.md'), encoding='utf-8') as f:
     long_description = f.read()
 
 setup(
-    name='pyandroidkx',
-    packages=find_packages(
-            include=["pyandroidkx", "pyandroidkx.*"]
-        ),
-    version='1.0.0',
-    description='A re-implementation of android java API in python with easy access to some Android functionality '
-                'like Notification,Reading of Contacts, accessing Webview Cookies, etc...',
+    name='kvdroid',
+    packages=['kvdroid'],
+    version='0.2.4',
+    description='Some Pyjnius tools for Kivy-Android developments',
     long_description=long_description,
     long_description_content_type='text/markdown',
-    author='Akubue-Izundu Kenechukwu',
-    author_email='kengoon19@gmail.com',
+    author='Yunus Ceyhan',
+    author_email='yunus.ceyhn@gmail.com',
     url='https://github.com/yunus-ceyhan/Kvdroid',
-    keywords=['Android', 'Androidx', 'Python', 'Kivy', 'KivyMD'],
-    install_requires=["pyjnius"],
+    keywords=['Android', 'Python', 'Kivy'],
+    install_requires=["kivy"],
     classifiers=[],
 )
